+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/api/NetIdService.kt
@@ -18,16 +18,20 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.util.Log
+import android.view.View
+import android.widget.Button
+import android.widget.LinearLayout
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
+import com.google.android.material.button.MaterialButton
+import de.netid.mobile.sdk.R
 import de.netid.mobile.sdk.appauth.AppAuthManager
 import de.netid.mobile.sdk.appauth.AppAuthManagerFactory
 import de.netid.mobile.sdk.appauth.AppAuthManagerListener
 import de.netid.mobile.sdk.model.*
 import de.netid.mobile.sdk.permission.PermissionManager
 import de.netid.mobile.sdk.permission.PermissionManagerListener
-import de.netid.mobile.sdk.ui.AuthorizationFragmentListener
-import de.netid.mobile.sdk.ui.AuthorizationLoginFragment
-import de.netid.mobile.sdk.ui.AuthorizationPermissionFragment
+import de.netid.mobile.sdk.ui.*
 import de.netid.mobile.sdk.userinfo.UserInfoManager
 import de.netid.mobile.sdk.userinfo.UserInfoManagerListener
 import de.netid.mobile.sdk.util.JsonUtil
@@ -39,6 +43,13 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
 
     private const val appIdentifierFilename = "netIdAppIdentifiers.json"
     private const val broker = "broker.netid.de"
+    private var layerStyle: NetIdLayerStyle = NetIdLayerStyle.Solid
+    private var buttonStyle: NetIdButtonStyle = NetIdButtonStyle.WhiteSolid
+    private var netIdLogoResource:Int = R.drawable.ic_netid_logo_button
+    private var buttonBackgroundResource:Int = R.color.authorization_agree_button_color
+    private var buttonForegroundResource:Int = R.color.authorization_agree_text_color
+    private var buttonOutlineResource:Int = R.color.authorization_close_button_color
+    private var buttonStrokeWidthResource:Int = R.dimen.authorization_close_button_stroke_width
 
     private var netIdConfig: NetIdConfig? = null
 
@@ -48,6 +59,12 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
 
     private val availableAppIdentifiers = mutableListOf<AppIdentifier>()
     private val netIdServiceListeners = mutableSetOf<NetIdServiceListener>()
+
+    private var permissionContinueButtonFragment:Fragment? = null
+    private var loginContinueButtonFragment:Fragment? = null
+    private var appButtonFragmentsForPermission = mutableMapOf<String, Fragment>()
+    private var appButtonFragmentsForLogin = mutableMapOf<String, Fragment>()
+    private var appButtonFragmentsForLoginPermission = mutableMapOf<String, Fragment>()
 
     fun addListener(listener: NetIdServiceListener) {
         netIdServiceListeners.add(listener)
@@ -65,14 +82,22 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
             }
 
             this.netIdConfig = netIdConfig
-            setupAuthManagerAndFetchConfiguration(context, broker)
+            setupAuthManagerAndFetchConfiguration(context)
             setupUserInfoManager()
             setupPermissionManager()
+            checkAvailableNetIdApplications(context)
         }
     }
 
-    fun getAuthorizationFragment(activity: Activity, authFlow: NetIdAuthFlow, forceApp2App: Boolean = false): Fragment? {
-        checkAvailableNetIdApplications(activity)
+    /**
+     * Gets the authorization fragment for a requested authorization flow.
+     * @param context Context to use.
+     * @param authFlow Authorization flow to use, see ``NetIdAuthFlow``.
+     * @param forceApp2App Set to true, if only app2app is allowed.
+     * @return Fragment for authorization.
+     */
+    fun getAuthorizationFragment(context: Context, authFlow: NetIdAuthFlow, forceApp2App: Boolean = false): Fragment? {
+        checkAvailableNetIdApplications(context)
         // If there are no ID apps installed, but forceApp2App is true, return with an error.
         if ((availableAppIdentifiers.isEmpty()) && forceApp2App) {
             this.onAuthorizationFailed(NetIdError(NetIdErrorProcess.Authentication, NetIdErrorCode.NoIdAppInstalled))
@@ -90,7 +115,7 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
                 config.claims,
                 effectivePrompt,
                 authFlow,
-                activity
+                context
             )?.let {
                 when (authFlow) {
                     NetIdAuthFlow.Login, NetIdAuthFlow.LoginPermission ->
@@ -107,6 +132,165 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
         return null
     }
 
+    /**
+     * Sets the style to use for all layers when using the layer flow.
+     * @param layerStyle button style to set, can be any of ``NetIdLayerStyle``, defaults to ``NetIdLayerStyle.Solid``
+     */
+    fun setLayerStyle(layerStyle: NetIdLayerStyle) {
+        this.layerStyle = layerStyle
+    }
+
+    /**
+     * Gets the currently set style that's used for all layers when using the layer flow.
+     * @return Currently set style
+     */
+    fun getLayerStyle(): NetIdLayerStyle {
+        return layerStyle
+    }
+
+    /**
+     * Sets the style to use for all buttons when using the button flow.
+     * @param buttonStyle button style to set, can be any of ``NetIdButtonStyle``, defaults to ``NetIdButtonStyle.GraySolid``
+     */
+    fun setButtonStyle(buttonStyle: NetIdButtonStyle, activity: Activity) {
+        this.buttonStyle = buttonStyle
+
+        if (permissionContinueButtonFragment != null) {
+            (permissionContinueButtonFragment as PermissionContinueButtonFragment).setButtonStyle(buttonStyle)
+        }
+
+        if (loginContinueButtonFragment != null) {
+            (loginContinueButtonFragment as LoginContinueButtonFragment).setButtonStyle(buttonStyle)
+        }
+
+        appButtonFragmentsForPermission.forEach {
+            val frag = it.value as AccountProviderAppButtonFragment
+            frag.setButtonStyle(buttonStyle)
+        }
+
+        appButtonFragmentsForLogin.forEach {
+            val frag = it.value as AccountProviderAppButtonFragment
+            frag.setButtonStyle(buttonStyle)
+        }
+
+        appButtonFragmentsForLoginPermission.forEach {
+            val frag = it.value as AccountProviderAppButtonFragment
+            frag.setButtonStyle(buttonStyle)
+        }
+    }
+
+    /**
+     * Gets the currently set style that's used for all buttons when using the button flow.
+     * @return Currently set style
+     */
+    fun getButtonStyle(): NetIdButtonStyle {
+        return buttonStyle
+    }
+
+    /**
+     * Returns the count of installed account provider apps.
+     * Use this function only if you intent to build your very own authorization dialog.
+     * @return: Count of installed account provider apps.
+     */
+    fun getCountOfAccountProviderApps(context: Context): Int {
+        checkAvailableNetIdApplications(context)
+        return availableAppIdentifiers.count()
+    }
+
+    /**
+     * Returns the keys of installed account provider apps. With these keys, you can request buttons for specific account provider apps identified by their key aka name.
+     * Use this function only if you intent to build your very own authorization dialog.
+     * @return: Array of keys of installed account provider apps.
+     */
+    fun getKeysForAccountProviderApps(): Array<String> {
+        val result = mutableListOf<String>()
+        availableAppIdentifiers.forEach {
+            result.add(it.name)
+        }
+        return result.toTypedArray()
+    }
+
+    /**
+     * Returns the authorization intent for a requested flow.
+     * @param flow Requested flow.
+     * @param context Context to use.
+     * @return Authorization intent.
+     */
+    fun authIntentForFlow(flow: NetIdAuthFlow, context: Context): Intent? {
+        netIdConfig?.let { config ->
+            return appAuthManager.getAuthorizationIntent(
+                config.clientId,
+                config.redirectUri,
+                config.claims,
+                config.promptWeb,
+                flow,
+                context
+            )
+        }
+        return null
+    }
+
+    /**
+     * Returns the continue button (as a fragment) in case of a permission flow dialog.
+     * Use this function only if you intent to build your very own authorization dialog.
+     * @param continueText Alternative text to set on the button. If empty, the default will be used.
+     * @return Fragment for authorization.
+     */
+    fun permissionContinueButtonFragment(continueText: String = ""): Fragment {
+        if (permissionContinueButtonFragment == null) {
+            permissionContinueButtonFragment = PermissionContinueButtonFragment(this, continueText)
+        }
+        return permissionContinueButtonFragment as Fragment
+    }
+
+    /**
+     * Returns the continue button (as a fragment) in case of a login flow dialog.
+     * Use this function only if you intent to build your very own authorization dialog.
+     * @param continueText Alternative text to set on the button. If empty, the default will be used.
+     * @param authFlow Must either be .Login or .LoginPermission. If is set to .Permission, an error will be thrown.
+     * @return Fragment for authorization.
+     */
+    fun loginContinueButtonFragment(continueText: String = "", flow: NetIdAuthFlow): Fragment {
+        if (loginContinueButtonFragment == null) {
+            loginContinueButtonFragment = LoginContinueButtonFragment(this, continueText, flow)
+        }
+        return loginContinueButtonFragment as Fragment
+    }
+
+    /**
+     * Returns the button for a certain account provider app for a requested ``NetIdAuthFlow``
+     * Use this function only if you intent to build your very own authorization dialog.
+     * @param key Key denoting one of the installed account provider apps. Use ``getKeysForAccountProviderApps`` first to get the keys/names of all installed account provider apps.
+     * @param authFlow Can be any of .Permission, .Login or .LoginPermission.
+     * @param continueText Alternative text to set on the button. If empty, the default will be used.
+     * @returns Button with text and label for the chosen id app. If index is out of bounds or no app is installed, returns an empty view.
+     */
+    fun accountProviderAppButtonFragment(key: String, flow: NetIdAuthFlow, continueText: String = ""): Fragment {
+        val keys = getKeysForAccountProviderApps()
+        val appButtonFragments = when (flow) {
+            NetIdAuthFlow.Permission -> appButtonFragmentsForPermission
+            NetIdAuthFlow.Login -> appButtonFragmentsForLogin
+            NetIdAuthFlow.LoginPermission -> appButtonFragmentsForLoginPermission
+        }
+        if (keys.contains(key)) {
+            return if (appButtonFragments.containsKey(key)) {
+                appButtonFragments[key] as Fragment
+            } else {
+                val index = keys.binarySearch(key)
+                val app = AccountProviderAppButtonFragment(this, availableAppIdentifiers[index], flow, continueText)
+                appButtonFragments[key] = app
+                app
+            }
+        }
+        throw ArrayIndexOutOfBoundsException()
+    }
+
+    /**
+     * Checks whether there is a network connection or not.
+     * @param context Context to use.
+     * @param process: In case of an error, denotes the process that the error is responsible for.
+     * @returns bool
+     */
     private fun handleConnection(context: Context, process: NetIdErrorProcess): Boolean {
         return if (ReachabilityUtil.hasConnection(context)) {
             true
@@ -118,6 +302,10 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
         }
     }
 
+    /**
+     * Fetches the user info.
+     * @param context Context to use.
+     */
     fun fetchUserInfo(context: Context) {
         if (handleConnection(context, NetIdErrorProcess.UserInfo)) {
             var error: NetIdError? = null
@@ -141,6 +329,11 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
         }
     }
 
+    /**
+     * Fetch permissions.
+     * @param context Context to use.
+     * @param collapseSyncId: Boolean value to indicate whether syncId is used or not.
+     */
     fun fetchPermissions(context: Context, collapseSyncId: Boolean = true) {
         if (handleConnection(context, NetIdErrorProcess.PermissionRead)) {
             var error: NetIdError? = null
@@ -157,6 +350,12 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
         }
     }
 
+    /**
+     * Update permissions.
+     * @param context Context to use.
+     * @param permission Permissions to set, of type ``NetIdPermissionUpdate``.
+     * @param collapseSyncId Boolean value to indicate if syncId is used or not.
+     */
     fun updatePermission(context: Context, permission: NetIdPermissionUpdate, collapseSyncId: Boolean = true) {
         if (handleConnection(context, NetIdErrorProcess.PermissionWrite)) {
             var error: NetIdError? = null
@@ -174,10 +373,23 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
         }
     }
 
-    private fun setupAuthManagerAndFetchConfiguration(context: Context, host: String) {
+    /**
+     * Function to end a session.
+     * The net ID service itself still remains initialized but all information about authorization/authentication is discarded.
+     * To start a new session, call ``authorize(destinationScheme:currentViewController:authFlow)`` again.
+     */
+    fun endSession() {
+        Log.i(javaClass.simpleName, "netId service did end session successfully")
+        appAuthManager.endSession()
+        for (item in netIdServiceListeners) {
+            item.onSessionEnd()
+        }
+    }
+
+    private fun setupAuthManagerAndFetchConfiguration(context: Context) {
         appAuthManager = AppAuthManagerFactory.createAppAuthManager(context)
         appAuthManager.listener = this
-        appAuthManager.fetchAuthorizationServiceConfiguration(host)
+        appAuthManager.fetchAuthorizationServiceConfiguration(broker)
     }
 
     private fun setupUserInfoManager() {
@@ -195,10 +407,6 @@ object NetIdService : AppAuthManagerListener, AuthorizationFragmentListener,
         availableAppIdentifiers.addAll(installedAppIdentifiers)
     }
 
-    fun endSession() {
-        Log.i(javaClass.simpleName, "netId service did end session successfully")
-        appAuthManager.endSession()
-    }
 
 // AppAuthManagerListener functions